mainapi/requirements.txt
black==19.3b0
pytest==5.2.1
pytest-cov==2.8.1
python-dotenv==0.10.3
Flask==2.0.1
responder==2.0.7
simple-rest-client==1.0.8